feat: add the navigation from home page to routines
