+++ apps/frontend/app/Routines/Assets/routineHero.webp
