+++ apps/frontend/app/Routines/Assets/yourProgress.webp
