+++ apps/frontend/app/Routines/Components/RoutineBenefits.tsx
