+++ apps/frontend/app/Routines/Components/RoutineBox.tsx
