+++ apps/frontend/app/Routines/Components/RoutineHero.tsx
